mysql: retain character set across autoreconnect
Character set settings are not retained across drop-reestablish
cycles (in a autoreconnect scenario). To that end, the function
mysql_set_character_set() is needed instead, which also implies
issuing SET NAMES to the server on every connection establishment.
(This seems to be behind the "CHARSET utf8" command that only the
MySQL CLI understands.)